requirements.txt
# requirements.txt - Python 3.12.3 Compatible Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
httpx>=0.25.0

# ML/DL
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging & Monitoring
tensorboard>=2.15.0
wandb>=0.16.0  # Optional: Weights & Biases

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# Note: PyTorch installed separately via --index-url for CPU/GPU compatibility